Remove book amnesia
One of crawl's most secret features, book amnesia allowed you to
destroy a book to forget one of its spells. In principle this is a
cool tradeoff (do you want to give up the option of ever learning
any of the spells in the book?), but it's rarely relevant in
practice (even for players who know about it!), and if it were
relevant, good play would involve a complex optimization game,
cross-referencing available books and the spells within. Which does
not seem especially fun to get dragged into midway through a
perfectly good game of DCSS.

?amnesia generation bumped up by ~12% to compensate.