Filter out conflicting books from xv
When examining a monster with multiple spellsets, don't display
spellsets that the player can deduce the monster can't have. (If
it's cast a spell that's not in that spellset while the player was
watching.)

I'd rather display them in darkgrey, but we don't currently support
color in monster spellset descriptions.