Cut each Hell from seven floors to four
The four Hells each have six floors empty of loot, with unique layouts
meant to establish the branch themes and denizens while regularly raining
down unique direct effects onto the player to emphasize the somewhat
unique dynamic of encouraged diving. While any sort of monster can slow
down players to support any numbers of effects, both lists are weighted
towards and filled with very ignorable results for people venturing into
an extended branch, and the repeated redundant lengths of each branch
betrays a reliance on randomization and patience more than sustainably
mean results (outside of the end floor monster distributions, of course).
This is unideal for postgame challenges, and exacerbates the redundant
length of extended.

As a result, this cuts the floors of each Hell from seven floors to four,
which should still promote some diving but not all. Oncoming commits will
use this somewhat shorter Hell context to rescale and revise both monster
sets and hell effects, for branch identity strengthening and to bring
a sense of consistent brutality to the arrangements.