Add a new species: Bultungin
Add the Bultungin, humanoids with the heads of dogs.

Bultungins are unique in their skill training process: instead of having
a static aptitude for each skill, it dynamically scales based on the
current skill level of the skill they are attempting to train. They
start out with an effective aptitude of +4 for all skills until skill
level 7 for each skill. After that, their aptitude for that skill
reduces by 2 for each skill level, with a floor of -6 at skill level 11.

This makes it more efficient for Bultungins to spread out their skills
over attempting to focus on raising a single skill to a high level.

It takes a Human 1400 experience to bring a skill from 0 to 7. Bultungin
need only 700 XP to get to that point. After 7, however, Humans begin to
surpass Bultungin: it takes them only 7050 total XP to advance a skill
to level 15 while a Bultungin takes 13693 total XP to get there, nearly
twice as much.

Bultungins also have see invisible, fangs 1, and can detect any items
within 7 tiles with their strong nose. Other attributes include base
stats of 7 STR/8 INT/9 DEX with increases of sid/5 levels, +2 MR/level,
and baseline XP/HP/MP values.